Fix more test sign mismatches
Fix some signed/unsigned comparison warnings in
viewer_draw_translator_test.cc and
rigid_body_tree_lcm_publisher_test.cc.

See also 0b696701ab06, 0854bc3a8186 and fe4790765811.

+++ drake/systems/plants/rigid_body_plant/test/viewer_draw_translator_test.cc
@@ -67,7 +67,7 @@ GTEST_TEST(ViewerDrawTranslatorTests, BasicTest) {
   double time = 0;
   std::vector<uint8_t> message_bytes;
   viewer_draw_translator.Serialize(time, generalized_state, &message_bytes);
-  EXPECT_GT(message_bytes.size(), 0);
+  EXPECT_GT(message_bytes.size(), 0u);
 
   // Verifies that the serialized message is correct. This entails:
   //     (1) manually creating a the correct `drake::lcmt_viewer_draw`
@@ -107,7 +107,7 @@ GTEST_TEST(ViewerDrawTranslatorTests, BasicTest) {
   expected_message.quaternion.push_back(zero_quaternion);
 
   const int byte_count = expected_message.getEncodedSize();
-  EXPECT_EQ(byte_count, message_bytes.size());
+  EXPECT_EQ(byte_count, static_cast<int>(message_bytes.size()));
 
   std::vector<uint8_t> expected_bytes(byte_count);
   expected_message.encode(expected_bytes.data(), 0, byte_count);
